+++ Cargo.toml
@@ -54,7 +54,7 @@ wasi-common = { workspace = true, default-features = true, features = ["exit"], 
 wasmtime-wasi = { workspace = true, default-features = true, optional = true }
 wasmtime-wasi-nn = { workspace = true, optional = true }
 wasmtime-wasi-threads = { workspace = true, optional = true }
-wasmtime-lind = { path = "crates/lind" }
+wasmtime-lind-multi-process = { path = "crates/lind-multi-process" }
 wasmtime-lind-common = { path = "crates/lind-common" }
 wasmtime-lind-utils = { path = "crates/lind-utils" }
 wasmtime-wasi-http = { workspace = true, optional = true }
@@ -200,7 +200,7 @@ wasmtime-wasi = { path = "crates/wasi", version = "23.0.0", default-features = f
 wasmtime-wasi-http = { path = "crates/wasi-http", version = "=23.0.0", default-features = false }
 wasmtime-wasi-nn = { path = "crates/wasi-nn", version = "23.0.0" }
 wasmtime-wasi-threads = { path = "crates/wasi-threads", version = "23.0.0" }
-wasmtime-lind = { path = "crates/lind", version = "23.0.0" }
+wasmtime-lind-multi-process = { path = "crates/lind-multi-process", version = "23.0.0" }
 wasmtime-lind-common = { path = "crates/lind-common", version = "23.0.0" }
 wasmtime-lind-utils = { path = "crates/lind-utils", version = "23.0.0" }
 wasmtime-component-util = { path = "crates/component-util", version = "=23.0.0" }
